backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-dotenv==1.1.0
opencv-python-headless==4.11.0.86
numpy==2.2.4
pillow==11.2.1
firebase-admin==6.8.0
google-genai==1.47.0
pytest==8.3.3
pytest-asyncio==0.24.0
